kill the player on quit if they were in combat

+++ handler/quit.go
@@ -1,10 +1,12 @@
 package handler
 
+import "github.com/df-mc/dragonfly/server/entity/damage"
+
 func (h PlayerHandler) HandleQuit() {
-	/*if _, combat := h.p.Combat(); combat {
+	if !h.p.CombatCD().Expired() {
 		if src, ok := h.p.LastHurt(); ok {
 			h.p.Kill(damage.SourceEntityAttack{Attacker: src})
 		}
-	}*/
+	}
 	h.p.Server().RemovePlayer(h.p)
 }
